backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.352
qdrant-client==1.7.0
pandas==2.0.3
requests==2.31.0
sentence-transformers==2.2.2
ollama==0.1.0
pydantic==2.4.2
torch==2.0.1
huggingface-hub==0.16.4
